+++ src/main/java/com/bringup/admin/notify/repository/NotificationRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
-    List<Notification> findByUserIdAndRoleAndIsReadFalse(Long userId, RolesType role);  // 읽지 않은 알림 조회
+    List<Notification> findByUserIdAndRoleAndIsReadFalse(int userId, RolesType role);  // 읽지 않은 알림 조회
 }
