+++ src/main/java/com/bringup/common/security/jwt/JwtProvider.java
@@ -104,7 +104,7 @@ public class JwtProvider {
 		}
 
 		UserDetails company = CompanyDetailsImpl.builder()
-			.id(claims.get("id", Long.class))
+			.id(claims.get("id", int.class))
 			.email(claims.getSubject())
 			.password(null)
 			.authorities(authorities)
